% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectPelagicVT.R
\name{connectMed}
\alias{connectMed}
\title{Create Connection to the Mediterranean Monitoring Database.
This connection grant permission to select on the sharkpulse table.}
\usage{
connectMed(dbuser, dbpass, db = "med_monitoring")
}
\arguments{
\item{dbuser}{role name for database}

\item{dbpass}{database password for role `dbuser`}

\item{db}{database password for role `dbuser`}
}
\description{
Create Connection to the Mediterranean Monitoring Database.
This connection grant permission to select on the sharkpulse table.
}
\examples{
con = connectMed()
dat
}
