% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectPelagicVT.R
\name{connectPelagic}
\alias{connectPelagic}
\title{Create Connection to the Pelagic Database on sharkpulse.cnre.vt.edu.}
\usage{
connectPelagic(dbuser, dbpass)
}
\arguments{
\item{dbuser}{role name for database}

\item{dbpass}{database password for role `dbuser`}
}
\description{
Create Connection to the Pelagic Database.
This connection grant permission to select on the sharkpulse table.
}
\examples{
con = connectPelagic()
dat = selectData(con, "select * from sharkpulse limit1")
dat
}
