% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfc.R
\name{dfc}
\alias{dfc}
\title{Return Distance from Coast}
\usage{
dfc(dat, sample = TRUE, nsamples = 50)
}
\arguments{
\item{dat}{dataset of points. this should not contain missing coordinates}

\item{sample}{whether a sample should be taken for calculating distance - this process takes approximately 2 min for 1000 points}

\item{nsamples}{number of samples to take}
}
\description{
Calculate distance from coast of a set of points
}
