% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFAOareas.R
\name{FAO_map}
\alias{FAO_map}
\title{Map species records on a FAO map}
\usage{
FAO_map(species, records = dat)
}
\arguments{
\item{species}{species name}

\item{records}{sharkPulse records}
}
\description{
Map species records on a FAO map
}
