% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2html.R
\name{flickr2html}
\alias{flickr2html}
\title{Create html table from flickr.search function}
\usage{
flickr2html(dat)
}
\arguments{
\item{dat}{data.frame of pre-validated pictures. There three columns: Species, ID, Img.}
}
\description{
Create html table from flickr.search function
}
