% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterFlickr.R
\name{flickrFilter}
\alias{flickrFilter}
\title{Filter flickr Results}
\usage{
flickrFilter(query = "shark", startUploadDate = "2015-01-02",
  endUploadDate = "2015-12-02", geoBox = "-180, -90, 180, 90")
}
\arguments{
\item{startUploadDate}{date string yyyy-mm-dd}

\item{endUploadDate}{date string yyyy-mm-dd}

\item{geobox}{character string for a geographic box "west, south, east, north"}
}
\description{
The function will crawl the Flickr database extract the images with the given tag  ("shark") and then filterm them excluding points on land.
}
