% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeInlands.R
\name{getBufferedWorld}
\alias{getBufferedWorld}
\title{Create a global buffer}
\usage{
getBufferedWorld(buffer_km)
}
\arguments{
\item{buffer_km}{distance in km to use as a bufer, negative values shrink the continents and positive values expand them.}
}
\description{
Takes a global polygon  apply a buffer of `buffer_km` and generate a unified buffered polygon. If `buffer_km` is positive it expand the continents, if `buffer_km` is negative, it shriks the continents. When shrinking the continents, the funtion prevents that internal corridors (gaps) are created because of shrinking national borders.
}
