% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterFlickr.R
\name{getFlickrImages}
\alias{getFlickrImages}
\title{Get Flickr Images}
\usage{
getFlickrImages(query, startUploadDate, endUploadDate, geoBox)
}
\arguments{
\item{query}{flickr tag to use to extract images}

\item{startUploadDate}{date string yyyy-mm-dd}

\item{endUploadDate}{date string yyyy-mm-dd}

\item{geobox}{character string for a geographic box "west, south, east, north"}
}
\value{
it writes a file \code{output.csv} in the \code{py} folder.
}
\description{
The function will connect to Flickr and extract image links with a given tag ("shark") and then filterm them excluding points on land.
}
