% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInstagram (1).R, R/getInstagram.R
\name{getInstagram}
\alias{getInstagram}
\title{Get the entire sharkPulse table}
\usage{
getInstagram(dbuser, dbpass)

getInstagram(dbuser, dbpass)
}
\arguments{
\item{dbuser}{user with privileges to access the instagram table}

\item{dbpass}{password for user with privileges to access the instagram table}
}
\description{
Get the entire sharkPulse table

Get the instagram table
}
