% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIUCNlist.R
\name{getIUCNlist}
\alias{getIUCNlist}
\title{Get the complete list of Chondrichthyes from the IUCN}
\usage{
getIUCNlist(
  key = "7532c24e6a4d09f24d595f4925beaaf65a9dd07e4ee31fb1c51dece2991df213"
)
}
\arguments{
\item{key}{is the IUCN Red List key}
}
\description{
It uses the rredlist package to compile the complete list of cartilaginous fishes in the IUCN Red List of Threatened species
}
