% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSharkPulseMonitor.R
\name{getSharkPulseMonitor}
\alias{getSharkPulseMonitor}
\title{Retrieve pulse monitor}
\usage{
getSharkPulseMonitor(dbuser, dbpass)
}
\arguments{
\item{dbuser}{sp user with privileges to access the sharkpulse table}

\item{dbpass}{sp password for user with privileges to access the sharkpulse table}
}
\description{
Retrieve pulse monitor
}
