% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFAOareas.R
\name{isinfao}
\alias{isinfao}
\title{Check if a point is within the vector of FAO areas indicated for the species}
\usage{
isinfao(fao_code, species_name, areas)
}
\description{
Check if a point is within the vector of FAO areas indicated for the species
}
