% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IUCN_map.R
\name{IUCN_map}
\alias{IUCN_map}
\title{Plot SP data over IUCN distribution map}
\usage{
IUCN_map(species, records = dat, shapes = shark_sh)
}
\arguments{
\item{species}{scientific name of the focal species}

\item{records}{usually sharkpulse data but can be any occurrence record dataset}

\item{shapes}{shaape files for plotting the expected distribution range}
}
\description{
Plot SP data over IUCN distribution map
}
