% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openImage.R
\name{openImage}
\alias{openImage}
\title{Open image from the sarkPulse archive}
\usage{
openImage(img_name, baselineUser = "francesco")
}
\arguments{
\item{img_name}{name of teh image to open}
}
\description{
this function require an image name as indicated in the sharkPulse table. Use selectData to get the image names of desired data.
}
