% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2html.R
\name{prevalidated2html}
\alias{prevalidated2html}
\title{Export Pre-validated Pictures to html}
\usage{
prevalidated2html(dat = dat, file = "validate.html")
}
\arguments{
\item{dat}{data.frame of pre-validated pictures. There three columns: Species, ID, Img.}
}
\description{
This function exports the pre-validated pictures in an html file for further screening
}
