% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeAqua.R
\name{removeAqua}
\alias{removeAqua}
\title{Remove records close to aquaria}
\usage{
removeAqua(aquaria = aqua, records = dat)
}
\arguments{
\item{aquaria}{location data of aquaria}

\item{records}{location data of records}
}
\description{
The function will filter records by removing those close to aquaria
}
