% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeAqua.R
\name{removeAqua2}
\alias{removeAqua2}
\title{Remove records close to aquaria}
\usage{
removeAqua2(aquaria, dat, max_distance_km = 1)
}
\arguments{
\item{aquaria}{location data of aquaria}

\item{max_distance_km}{distance in kilometers}

\item{records}{location data of records}
}
\description{
The function will filter records by removing those close to aquaria
}
