% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeAqua.R
\name{removeAquaPar}
\alias{removeAquaPar}
\title{Remove records close to aquaria - parellelized function}
\usage{
removeAquaPar(aquaria = aqua, records = dat, max_distance_km = 1)
}
\arguments{
\item{aquaria}{location data of aquaria}

\item{records}{location data of records}

\item{max_distance_km}{distance in kilometers}
}
\description{
The function will filter records by removing those close to aquaria
}
