% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeInlands.R
\name{removeInlands3}
\alias{removeInlands3}
\title{Remove points inland}
\usage{
removeInlands3(dat, buffer_km)
}
\arguments{
\item{dat}{dataset of locational points}

\item{buffer_km}{distance in km to use as a bufer, negative values shrink the continents and positive values expand them}
}
\description{
Takes a dataset of locational points (it needs to have latitude and longitude columns), and create a new variable classifying these points in inlands==TRUE if they are falling on land, and FALSE if they are falling in the ocean
}
