% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeNames.R
\name{sanitizeNames}
\alias{sanitizeNames}
\title{Sanitize scientific names}
\usage{
sanitizeNames(names, refdata = tx)
}
\arguments{
\item{names}{species names to sanitize}

\item{refdata}{data with reference species list}
}
\description{
convert a list of scientific names into their proper names, cross checking them with a validated list of official names. The function calculates the similarity between the word to check and the words in the reference list and takes teh closest match.
}
