% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeTaxonomy.R
\name{sanitizeTaxonomy}
\alias{sanitizeTaxonomy}
\title{Sanitize sharh names nomenclature}
\usage{
sanitizeTaxonomy(dat)
}
\arguments{
\item{dat}{data table with `species_name` column to sanitize}
}
\description{
convert a list of names containing species names with various frmatting error in a proper list of scientific names
}
