% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectData.R
\name{selectData}
\alias{selectData}
\title{Select Data from Database}
\usage{
selectData(con, statement)
}
\arguments{
\item{con}{connection to database}

\item{statement}{postgreSQL statement}
}
\description{
Select Data from Database
}
