% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSharkPulse.R
\name{subSharkPulse}
\alias{subSharkPulse}
\title{Substitute the entire sharkPulse table}
\usage{
subSharkPulse(dbuser, dbpass, csvfile)
}
\arguments{
\item{dbuser}{sCPUEdb user with privileges to access the sharkpulse table}

\item{dbpass}{sCPUEdb password for user with privileges to access the sharkpulse table}
}
\description{
Substitute the entire sharkPulse table
}
