% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeGS.R
\name{writeGS}
\alias{writeGS}
\title{Write a google sheet and render images with the IMAGE() formula}
\usage{
writeGS(df, sheet_name)
}
\arguments{
\item{df}{data frame to supply}

\item{sheet_name}{name of the google sheet}
}
\description{
Create a google sheet from a data frame
}
