% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTableInfo.R
\name{writeTableInfo}
\alias{writeTableInfo}
\title{Write info of a specific database table into a csv file}
\usage{
writeTableInfo(con = con, table_name = "instagram")
}
\arguments{
\item{con}{database connection}

\item{table_name}{nameof the specific table}
}
\description{
This is done to edit the file and include column descriptions. The resulting file will be named [table_name]Info.csv
}
