% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFAOareas.R
\name{getFaoNames}
\alias{getFaoNames}
\title{Get FAO area for records}
\usage{
getFaoNames(dat)
}
\arguments{
\item{dat}{usually sharkpulse data with latitude and longitude. It can be any occurrence record dataset.}
}
\description{
Function to get the FAO zone names in English for a set of points.
}
