% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFlickrImages.R
\name{getFlickrImages}
\alias{getFlickrImages}
\title{Get Flickr Images}
\usage{
getFlickrImages(text, geoBox = "-180,-90,180,90")
}
\arguments{
\item{text}{flickr tag to use to extract images}

\item{geobox}{character string for a geographic box "west, south, east, north"}
}
\value{
it writes a file \code{output.csv} in the \code{py} folder.
}
\description{
The function will connect to Flickr and extract image links with a given tag ("shark") and then filterm them excluding points on land.
}
