% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInstagram.R
\name{getInstagram}
\alias{getInstagram}
\title{Get the instagram table}
\usage{
getInstagram(dbuser, dbpass)
}
\arguments{
\item{dbuser}{user with privileges to access the instagram table}

\item{dbpass}{password for user with privileges to access the instagram table}
}
\description{
Get the instagram table
}
