% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSharkPulse.R
\name{getSharkPulse}
\alias{getSharkPulse}
\title{Get the entire sharkPulse table}
\usage{
getSharkPulse(dbuser, dbpass)
}
\arguments{
\item{dbuser}{sCPUEdb user with privileges to access the sharkpulse table}

\item{dbpass}{sCPUEdb password for user with privileges to access the sharkpulse table}
}
\description{
Get the entire sharkPulse table
}
