% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openImage.R
\name{openImage}
\alias{openImage}
\title{Open image from the sarkPulse archive}
\usage{
openImage(img_name, srvUser = "francesco")
}
\arguments{
\item{img_name}{name of the image to open}

\item{srvUser}{username in the remote server}
}
\description{
this function require an image name as indicated in the sharkPulse table and the training table. Use selectData to get the image names of desired data.
To work, you need to be able to access the remote server without password
}
